requirements.txt
flask==2.0.1
flask-restx==0.5.1
torch==1.9.0
torchvision==0.10.0
pillow==8.3.1
numpy==1.21.0
opencv-python==4.5.3.56
gunicorn==20.1.0
python-dotenv==0.19.0
werkzeug==2.0.1
requests

# For GradCAM visualization
opencv-python>=4.5.0
numpy>=1.19.0
torch>=1.7.0
torchvision>=0.8.0
Pillow>=8.0.0